+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20140910190904) do
+ActiveRecord::Schema.define(version: 20140919202446) do
 
   create_table "app_categories", force: true do |t|
     t.string   "name"
@@ -27,6 +27,12 @@ ActiveRecord::Schema.define(version: 20140910190904) do
     t.string "name"
     t.string "from"
     t.text   "documentation"
+  end
+
+  create_table "deployment_targets", force: true do |t|
+    t.string "name"
+    t.text   "cert_file"
+    t.text   "auth_blob"
   end
 
   create_table "images", force: true do |t|